+++ CHANGELOG.md
@@ -47,6 +47,7 @@
         - enclave (#176)
     - Reward worker. (#203)
     - Reward app provider. (#204)
+    - Reward dataset provider. (#207)
     - Handle callback. (#183, #208)
 - Implement `claimBoost`:
     - Refund requester. (#198, #201)